Remove cycles in module dependencies
Having cycles in the modules occaisionally prevents the extension from starting.

- Run GitHub Actions check for cycles in module dependencies using dpdm
- Fix cli -> tools cycle
- Remove cluster -> clusterViewLoader cycle
- Move name validation functions out of OpenShiftItem to avoid cycles
- Refactor ServerlessFunction -> ServerlessFunctionModel to avoid cycles
- Remove view -> loader -> functions cycle
