+++ package.json
@@ -63,6 +63,7 @@
 		"lint": "eslint . --ext .ts --quiet",
 		"lint-fix": "eslint . --ext .ts --fix",
 		"lint-nic": "eslint . --ext .ts --no-inline-config",
+		"check-cycles": "dpdm --no-tree --no-warning --exit-code circular:1 ./src/extension.ts",
 		"bundle-tools": "node ./out/build/bundle-tools.js --platform",
 		"todo": "leasot **/*.ts --ignore node_modules -x",
 		"test": "npm run vscode:prepublish && node ./out/build/install-vscode.js && node ./out/build/run-tests.js unit",
@@ -140,6 +141,7 @@
 		"chokidar": "^3.5.3",
 		"codecov": "^3.8.2",
 		"esbuild": "^0.19.2",
+		"dpdm": "^3.13.1",
 		"esbuild-plugin-svgr": "^2.0.0",
 		"esbuild-sass-plugin": "^2.14.0",
 		"eslint": "8.48.0",
