+++ src/odo3.ts
@@ -5,12 +5,13 @@
 
 import { KubernetesObject } from '@kubernetes/client-node';
 import { CommandText } from './base/command';
-import { CliChannel, CliExitData } from './cli';
+import { CliChannel } from './cli';
 import { Command as CommonCommand, loadItems } from './k8s/common';
 import { Command as DeploymentCommand } from './k8s/deployment';
 import { DeploymentConfig } from './k8s/deploymentConfig';
 import { Command } from './odo/command';
 import { ComponentDescription } from './odo/componentTypeDescription';
+import { CliExitData } from './util/childProcessUtil';
 
 export interface Odo3 {
     getNamespaces(): Promise<KubernetesObject[]>;
