+++ test/unit/openshift/nameValidator.test.ts
@@ -6,44 +6,42 @@
 import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
 import { OdoImpl } from '../../../src/odo';
-import OpenShiftItem from '../../../src/openshift/openshiftItem';
+import * as NameValidator from '../../../src/openshift/nameValidator';
 import { wait } from '../../../src/util/async';
 import sinon = require('sinon');
 
-const {expect} = chai;
+const { expect } = chai;
 chai.use(sinonChai);
 
-suite('OpenShiftItem', () => {
-
+suite('nameValidator', function () {
     let sandbox: sinon.SinonSandbox;
 
-    setup(() => {
+    setup(function () {
         sandbox = sinon.createSandbox();
         sandbox.stub(OdoImpl.prototype, 'getActiveCluster').resolves('cluster');
     });
 
-    teardown(() => {
+    teardown(function () {
         sandbox.restore();
     });
 
-    test('Wait eventually exits', async () => {
+    test('Wait eventually exits', async function () {
         return wait();
     });
 
-    suite('validateMatches', ()=> {
-
-        test('returns validation message if provided value is not in lower case alphanumeric characters or "-"', ()=> {
-            const message = 'Not a valid Application name. Please use lower case alphanumeric characters or "-", start with an alphabetic character, and end with an alphanumeric character';
-            let appNames = OpenShiftItem.validateMatches(message, 'Nodejs-app');
+    suite('validateMatches', function () {
+        test('returns validation message if provided value is not in lower case alphanumeric characters or "-"', function () {
+            const message =
+                'Not a valid Application name. Please use lower case alphanumeric characters or "-", start with an alphabetic character, and end with an alphanumeric character';
+            let appNames = NameValidator.validateMatches(message, 'Nodejs-app');
             expect(appNames).equals(message);
-            appNames = OpenShiftItem.validateMatches(message, '2nodejs-app');
+            appNames = NameValidator.validateMatches(message, '2nodejs-app');
             expect(appNames).equals(message);
         });
 
-        test('returns undefined if provided value is in lower case alphanumeric characters', ()=> {
-            const validateMatches = OpenShiftItem.validateMatches(undefined, 'nodejs-app');
+        test('returns undefined if provided value is in lower case alphanumeric characters', function () {
+            const validateMatches = NameValidator.validateMatches(undefined, 'nodejs-app');
             expect(validateMatches).equals(null);
         });
     });
-
 });
